chore: clean e2e tests
