+++ engine/e2e-test/main.py
@@ -2,27 +2,16 @@ import pytest
 import sys
 
 ### e2e tests are expensive, have to keep engines tests in order
-from test_api_engine_list import TestApiEngineList
-from test_api_engine_install import TestApiEngineInstall
-from test_api_engine_get import TestApiEngineGet
-
-### models, keeps in order, note that we only uninstall engine after finishing all models test
-from test_api_model_pull_direct_url import TestApiModelPullDirectUrl
-from test_api_model_start_stop import TestApiModelStartStop
-from test_api_model_get import TestApiModelGet
-from test_api_model_list import TestApiModelList
-from test_api_model_update import TestApiModelUpdate
-from test_api_model_delete import TestApiModelDelete
+from test_api_engine import TestApiEngine
+from test_api_model import TestApiModel
 from test_api_model_import import TestApiModelImport
-from test_api_engine_uninstall import TestApiEngineUninstall
 
 ###
 from test_cli_engine_get import TestCliEngineGet
 from test_cli_engine_install import TestCliEngineInstall
 from test_cli_engine_list import TestCliEngineList
 from test_cli_engine_uninstall import TestCliEngineUninstall
-from test_cli_model_delete import TestCliModelDelete
-from test_cli_model_pull_direct_url import TestCliModelPullDirectUrl
+from test_cli_model import TestCliModel
 from test_cli_server_start import TestCliServerStart
 from test_cortex_update import TestCortexUpdate
 from test_create_log_folder import TestCreateLogFolder
