+++ engine/e2e-test/test_api_engine.py
@@ -1,29 +1,54 @@
-import time
-
 import pytest
 import requests
+import time
 from test_runner import (
-    run,
-    start_server_if_needed,
+    start_server,
     stop_server,
     wait_for_websocket_download_success_event,
 )
 
-
-class TestApiEngineUninstall:
+class TestApiEngine:
 
     @pytest.fixture(autouse=True)
     def setup_and_teardown(self):
         # Setup
-        start_server_if_needed()
+        success = start_server()
+        if not success:
+            raise Exception("Failed to start server")
 
         yield
 
         # Teardown
         stop_server()
 
+    # engines list
+    def test_engines_list_api_run_successfully(self):
+        response = requests.get("http://localhost:3928/engines")
+        assert response.status_code == 200
+    
+    # engines get
+    def test_engines_get_llamacpp_should_be_successful(self):
+        response = requests.get("http://localhost:3928/engines/llama-cpp")
+        assert response.status_code == 200
+        
+    # engines install
+    def test_engines_install_llamacpp_specific_version_and_variant(self):
+        data = {"version": "v0.1.35-27.10.24", "variant": "linux-amd64-avx-cuda-11-7"}
+        response = requests.post(
+            "http://localhost:3928/v1/engines/llama-cpp/install", json=data
+        )
+        assert response.status_code == 200
+
+    def test_engines_install_llamacpp_specific_version_and_null_variant(self):
+        data = {"version": "v0.1.35-27.10.24"}
+        response = requests.post(
+            "http://localhost:3928/v1/engines/llama-cpp/install", json=data
+        )
+        assert response.status_code == 200
+    
+    # engines uninstall
     @pytest.mark.asyncio
-    async def test_engines_uninstall_llamacpp_should_be_successful(self):
+    async def test_engines_install_uninstall_llamacpp_should_be_successful(self):
         response = requests.post("http://localhost:3928/v1/engines/llama-cpp/install")
         assert response.status_code == 200
         await wait_for_websocket_download_success_event(timeout=None)
@@ -33,7 +58,7 @@ class TestApiEngineUninstall:
         assert response.status_code == 200
 
     @pytest.mark.asyncio
-    async def test_engines_uninstall_llamacpp_with_only_version_should_be_failed(self):
+    async def test_engines_install_uninstall_llamacpp_with_only_version_should_be_failed(self):
         # install first
         data = {"variant": "mac-arm64"}
         install_response = requests.post(
@@ -50,7 +75,7 @@ class TestApiEngineUninstall:
         assert response.json()["message"] == "No variant provided"
 
     @pytest.mark.asyncio
-    async def test_engines_uninstall_llamacpp_with_variant_should_be_successful(self):
+    async def test_engines_install_uninstall_llamacpp_with_variant_should_be_successful(self):
         # install first
         data = {"variant": "mac-arm64"}
         install_response = requests.post(
@@ -62,7 +87,7 @@ class TestApiEngineUninstall:
         response = requests.delete("http://127.0.0.1:3928/v1/engines/llama-cpp/install")
         assert response.status_code == 200
 
-    def test_engines_uninstall_llamacpp_with_specific_variant_and_version_should_be_successful(
+    def test_engines_install_uninstall_llamacpp_with_specific_variant_and_version_should_be_successful(
         self,
     ):
         data = {"variant": "mac-arm64", "version": "v0.1.35"}
@@ -76,3 +101,5 @@ class TestApiEngineUninstall:
             "http://localhost:3928/v1/engines/llama-cpp/install", json=data
         )
         assert response.status_code == 200
+
+
